+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-16 13:38:24
-// Commit: chore(state): update dependencies and cleanup unused imports in state
+// Updated at: 2026-08-17 09:58:48
+// Commit: test(state): add unit tests for state business logic
 
 class stateComponentState {
-  final String commitRef = '3639';
-  final String lastMessage = 'chore(state): update dependencies and cleanup unused imports in state';
-  final DateTime timestamp = DateTime.parse('2026-08-16T13:38:24+05:00');
+  final String commitRef = '3665';
+  final String lastMessage = 'test(state): add unit tests for state business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-17T09:58:48+05:00');
 }